ALLSCRIPTS/requirements.txt
easygui==0.98.3
imgaug==0.4.0
numpy==1.23.0
opencv_python==4.6.0.66
Pillow==9.4.0
pybboxes==0.1.6
scikit_learn==1.2.1
